+++ modules/actor/cpr-demon.js
@@ -1,6 +1,7 @@
 /* globals Actor */
 import * as CPRRolls from "../rolls/cpr-rolls.js";
 import CPR from "../system/config.js";
+import CPRChat from "../chat/cpr-chat.js";
 import LOGGER from "../utils/cpr-logger.js";
 import SystemUtils from "../utils/cpr-systemUtils.js";
 
@@ -42,5 +43,18 @@ export default class CPRDemonActor extends Actor {
     const niceStatName = SystemUtils.Localize(CPR.demonStatList[statName]);
     const statValue = parseInt(this.data.data.stats[statName], 10);
     return new CPRRolls.CPRStatRoll(niceStatName, statValue);
+  }
+
+  /**
+   * Apply damage to the rez of the deamon.
+   * @param {int} damage - direct damage dealt
+   * @param {int} bonusDamage - bonus damage dealt
+   */
+  async _applyDamage(damage, bonusDamage) {
+    LOGGER.trace("_applyDamage | CPRDemonActor | Called.");
+    // As a deamon does not have any armor the damage will be simply subtracted from the REZ.
+    const currentRez = this.data.data.stats.rez.value;
+    await this.update({ "data.stats.rez.value": currentRez - damage - bonusDamage });
+    CPRChat.RenderDamageApplicationCard({ name: this.name, hpReduction: damage + bonusDamage, rezReduction: true });
   }
 }